Move hreflang and canonical middleware to shared
